requirements.txt
flask
flask-cors
azure.cosmos
python-dotenv